Updated the Services Page
Made major changes to the services page in order to improve the overall aesthetic and user experience. This includes but is not limited to adjusting the colour scheme to match that of the website's other pages, incorporating the standard buttons, changing the panel border properties as well as minor spelling error corrections.